+++ src/main/java/carmine/pittella/home/model/dto/MovimentoDto.java
@@ -3,7 +3,6 @@ package carmine.pittella.home.model.dto;
 import carmine.pittella.home.model.enums.TipologiaEnum;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,7 +16,7 @@ public class MovimentoDto {
     public LocalDateTime data;
     public TipologiaEnum tipologia;
     public String titolo;
-    public BigDecimal importo;
+    public Double importo;
     public String commento;
     public ContoDto conto;
     public CategoriaDto categoria;
