+++ src/main/java/carmine/pittella/home/model/dto/response/DashboardStatsResponseDto.java
@@ -5,16 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class DashboardStatsResponseDto {
 
-    private BigDecimal entrate;
-    private BigDecimal uscite;
-    private BigDecimal saldo;
+    private Double entrate;
+    private Double uscite;
+    private Double saldo;
 
 }
